Struct trackball::Frame

pub struct Frame<N: Copy + RealField> { /* private fields */ }

Frame wrt camera eye and target.

Implementations

impl<N: Copy + RealField> Frame<N>

pub fn look_at(target: Point3<N>, eye: &Point3<N>, up: &Vector3<N>) -> Self

Sets eye position inclusive its roll attitude and target position in world space.

pub fn eye(&self) -> Point3<N>

Eye position in world space.

pub fn set_eye(&mut self, eye: &Point3<N>, up: &Vector3<N>)

Sets eye position inclusive its roll attitude in world space preserving target position.

pub const fn target(&self) -> &Point3<N>

Target position in world space.

pub fn set_target(&mut self, target: Point3<N>)

Sets target position in world space preserving eye position inclusive its roll attitude.

pub const fn distance(&self) -> N

Distance between eye and target.

pub fn set_distance(&mut self, zat: N)

Sets distance between eye and target preserving target position.

pub fn scale(&mut self, rat: N)

Scales distance between eye and target by ratio preserving target position.

pub fn local_scale_around(&mut self, rat: N, pos: &Point3<N>)

Scales distance between eye and point in camera space by ratio preserving target position.

pub fn scale_around(&mut self, rat: N, pos: &Point3<N>)

Scales distance between eye and point in world space by ratio preserving target position.

pub fn local_slide(&mut self, vec: &Vector3<N>)

Slides camera eye and target by vector in camera space.

pub fn slide(&mut self, vec: &Vector3<N>)

Slides camera eye and target by vector in world space.

pub fn local_orbit(&mut self, rot: &UnitQuaternion<N>)

Orbits eye by rotation in camera space around target.

pub fn local_orbit_around(&mut self, rot: &UnitQuaternion<N>, pos: &Point3<N>)

Orbits eye by rotation in camera space around point in camera space.

pub fn orbit(&mut self, rot: &UnitQuaternion<N>)

Orbits eye by rotation in world space around target.

pub fn orbit_around(&mut self, rot: &UnitQuaternion<N>, pos: &Point3<N>)

Orbits eye by rotation in world space around point in world space.

pub fn look_around(&mut self, pitch: N, yaw: N, yaw_axis: &Unit<Vector3<N>>)

Orbits target around eye by pitch and yaw preserving roll attitude aka first person view.

Use fixed Self::yaw_axis() by capturing it when entering first person view.

pub fn local_pitch_axis(&self) -> Unit<Vector3<N>>

Positive x-axis in camera space pointing from left to right.

pub fn local_yaw_axis(&self) -> Unit<Vector3<N>>

Positive y-axis in camera space pointing from bottom to top.

pub fn local_roll_axis(&self) -> Unit<Vector3<N>>

Positive z-axis in camera space pointing from back to front.

pub fn pitch_axis(&self) -> Unit<Vector3<N>>

Positive x-axis in world space pointing from left to right.

pub fn yaw_axis(&self) -> Unit<Vector3<N>>

Positive y-axis in world space pointing from bottom to top.

pub fn roll_axis(&self) -> Unit<Vector3<N>>

Positive z-axis in world space pointing from back to front.

pub fn try_lerp_slerp(&self, other: &Self, t: N, epsilon: N) -> Option<Self>

Attempts to interpolate between two frames using linear interpolation for the translation part, and spherical linear interpolation for the rotation part.

Returns None if the angle between both rotations is 180 degrees (in which case the interpolation is not well-defined).

Arguments

self: The initial frame to interpolate from.

other: The final frame to interpolate toward.

t: The interpolation parameter between 0 and 1.

epsilon: The value below which the sinus of the angle separating both quaternion must be to return None.

pub fn renormalize(&mut self) -> N

Renormalizes eye rotation and returns its norm.

pub fn view(&self) -> Isometry3<N>

View transformation from camera to world space.

pub fn inverse_view(&self) -> Isometry3<N>

Inverse view transformation from world to camera space.

Uses less computations than Self::view().inverse().

pub fn cast<M: Copy + RealField>(self) -> Frame<M>where N: SubsetOf<M>,

Casts components to another type, e.g., between f32 and f64.