Struct trackball::Image

pub struct Image<N: Copy + RealField> { /* private fields */ }

Image as projection of Scope wrt Frame.

Implementations

impl<N: Copy + RealField> Image<N>

pub fn new(frame: &Frame<N>, scope: &Scope<N>, max: Point2<N>) -> Self

Computes initial transformations from frame, scope, and screen’s width and height.

pub fn compute(&mut self, frame: Frame<N>, scope: Scope<N>) -> Option<bool>

Recomputes only cached matrices whose parameters have changed, see Self::set_compute().

Returns Some(true) on success, Some(false) on failure, and None with no changes.

pub fn set_compute(&mut self, compute_mat: bool, compute_inv: bool)

Sets whether to compute transformation and inverse transformation with Self::compute().

Default is (true, true).

pub const fn pos(&self) -> &Point2<N>

Current position in screen space of hovering input or pointing device.

pub fn set_pos(&mut self, pos: Point2<N>)

Sets current position in screen space of hovering input or pointing device.

pub const fn max(&self) -> &Point2<N>

Maximum position in screen space as screen’s width and height.

pub fn set_max(&mut self, max: Point2<N>)

Sets maximum position in screen space as screen’s width and height.

pub const fn upp(&self) -> N

Cached unit per pixel on focus plane to scale/project positions/vectors onto focus plane.

pub const fn view_isometry(&self) -> &Isometry3<N>

Cached view isometry.

pub const fn view(&self) -> &Matrix4<N>

Cached view matrix.

pub fn compute_view(&mut self, frame: &Frame<N>)

Computes view isometry and matrix from frame wrt camera eye and target.

pub const fn projection(&self) -> &Matrix4<N>

Cached projection matrix.

pub fn compute_projection_and_upp(&mut self, zat: N, scope: &Scope<N>)

Computes projection matrix and unit per pixel on focus plane.

pub const fn transformation(&self) -> &Matrix4<N>

Cached projection view matrix.

pub fn compute_transformation(&mut self)

Computes projection view matrix.

pub const fn inverse_transformation(&self) -> &Matrix4<N>

Cached inverse projection view matrix.

pub fn compute_inverse_transformation(&mut self) -> bool

Computes inverse of projection view matrix.

Returns true on success.

pub fn clamp_pos_wrt_max(pos: &Point2<N>, max: &Point2<N>) -> Point2<N>

Clamps position in screen space wrt its maximum in screen space.

pub fn clamp_pos(&self, pos: &Point2<N>) -> Point2<N>

Clamps position in screen space.

pub fn transform_pos_and_max_wrt_max( pos: &Point2<N>, max: &Point2<N> ) -> (Point2<N>, Point2<N>)

Transforms position and its maximum from screen to camera space wrt its maximum.

pub fn transform_pos(&self, pos: &Point2<N>) -> Point2<N>

Transforms position from screen to camera space.

pub fn transform_vec(pos: &Vector2<N>) -> Vector2<N>

Transforms vector from screen to camera space.

pub fn project_pos(&self, pos: &Point2<N>) -> Point3<N>

Transforms position from screen to camera space and projects it onto focus plane.

pub fn project_vec(&self, vec: &Vector2<N>) -> Vector3<N>

Transforms vector from screen to camera space and projects it onto focus plane.

pub fn cast<M: Copy + RealField>(self) -> Image<M>where N: SubsetOf<M>,

Casts components to another type, e.g., between f32 and f64.