Struct trackball::Plane

#[repr(C)]
pub struct Plane<N: Copy + RealField> {
    pub normal: Unit<Vector3<N>>,
    pub bias: N,
}

Plane encoding position with singed bias along unit normal.

Realizes plane equation a*x+b*y+c*z+d=0 with unit normal [x, y, z] and signed bias d.

Fields
normal: Unit<Vector3<N>>

Plane unit normal.

bias: N

Signed bias along unit normal.

Implementations

impl<N: Copy + RealField> Plane<N>

pub fn new(normal: Unit<Vector3<N>>, distance: N) -> Self

Plane from unit normal and signed distance from the origin.

use trackball::{
    nalgebra::{Point3, Vector3},
    Plane,
};

// Plane intersecting y-axis at `5.0`.
let plane = Plane::new(Vector3::y_axis(), 5.0);

// Origin projected onto plane where plane intersects y-axis.
let point = plane.project_point(&Point3::origin());
assert_eq!(point, Point3::new(0.0, 5.0, 0.0));

// Bias is negated distance.
assert_eq!(plane.distance(), 5.0);
assert_eq!(plane.bias, -5.0);

// Plane intersecting y-axis at `-5.0`.
let plane = Plane::new(Vector3::y_axis(), -5.0);

// Origin projected onto plane where plane intersects y-axis.
let point = plane.project_point(&Point3::origin());
assert_eq!(point, Point3::new(0.0, -5.0, 0.0));

// Bias is negated distance.
assert_eq!(plane.distance(), -5.0);
assert_eq!(plane.bias, 5.0);

pub fn with_point(normal: Unit<Vector3<N>>, point: &Point3<N>) -> Self

Plane from unit normal with point in plane.

pub fn distance(&self) -> N

Signed orthogonal distance from the origin.

pub fn distance_from(&self, point: &Point3<N>) -> N

Signed orthogonal distance from point.

pub fn project_point(&self, point: &Point3<N>) -> Point3<N>

Projects point onto plane.

pub fn project_axis(&self, axis: &Unit<Vector3<N>>) -> Unit<Vector3<N>>

Projects axis onto plane.

pub fn project_vector(&self, vector: &Vector3<N>) -> Vector3<N>

Projects vector onto plane.

pub fn angle_between(&self, a: &Vector3<N>, b: &Vector3<N>) -> N

Singed angle from a to b where both vectors are in the plane.

pub fn rotate_by(self, rot: &UnitQuaternion<N>) -> Self

Rotates plane.

pub fn translate_by(self, vec: &Vector3<N>) -> Self

Translates plane.

pub fn transform_by(self, iso: &Isometry3<N>) -> Self

Transforms plane by direct isometry, i.e., rotation followed by translation.

use core::f64::{
    consts::FRAC_PI_2,
    EPSILON,
};
use trackball::{
    approx::AbsDiffEq,
    nalgebra::{Isometry3, Vector3},
    Plane,
};

// Plane intersecting y-axis at `5.0`.
let plane = Plane::new(Vector3::y_axis(), 5.0)
    .transform_by(&Isometry3::new(
        // Translation by after rotation.
        Vector3::new(-5.0, 0.0, 0.0),
        // Rotation by 90 degrees before translation.
        Vector3::new(0.0, 0.0, FRAC_PI_2),
    ));
// Plane intersecting x-axis at `-10.0`.
assert!(plane.abs_diff_eq(&Plane::new(-Vector3::x_axis(), 10.0), EPSILON));

pub fn cast<M: Copy + RealField>(self) -> Plane<M>where N: SubsetOf<M>,

Casts components to another type, e.g., between f32 and f64.